refactor: isolate table cell mutations
